+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix one more invalid error with colocated directories
 - Revert "Recognize links starting with `www` as external for the link checker" as they won't be external links in practice
 - Use page.summary for atom.xml if available
+- Fix cachebusting not working with binary files
 
 ## 0.17.1 (2023-02-24)